+++ components/organisms/header.vue
@@ -1,6 +1,6 @@
 <template>
   <nuxt-link :to="{name:'index'}">
-    <img src="/menu_logo_navbar.png">
+    <img src="/images/menu_logo_navbar.png">
     <span class="md-title">MenuApp</span>
   </nuxt-link>
 </template>