+++ pages/menu/register.vue
@@ -20,6 +20,7 @@
 
 <script>
 export default {
+  middleware: 'authenticated',
   data () {
     return {
       menu: {
@@ -34,6 +35,16 @@ export default {
       sending: false
     }
   },
+  created () {
+    // check if there is a menu in the params
+    // if there is a menu in the params use that menu to edit
+    const chosenMenu = this.$store.getters.getSelectedMenu
+    if (!Object.keys(chosenMenu).length) {
+      this.$store.commit('setSelectedMenu', this.menu)
+    } else {
+      this.menu = chosenMenu
+    }
+  },
   methods: {
     /**
      * add menu to firebase then redirect to list page
@@ -43,7 +54,7 @@ export default {
       if (this.validateMenu() && this.sending === false) {
         this.sending = true
         // add to firebase
-        this.$store.dispatch('menus/writeMenuToFirebase', this.menu)
+        this.$store.dispatch('writeMenuToFirebase')
           .then(() => {
             this.$router.push({ name: 'list' })
           })