+++ pages/menu/register/info/index.vue
@@ -1,9 +1,10 @@
 <template>
-  <editInfo :menu="menu" @setMenuTitle="setMenuTitle" @setMenuDescription="setMenuDescription" />
+  <editInfo :menu="menu" />
 </template>
 
 <script>
 export default {
+  middleware: 'authenticated',
   components: {
     editInfo: () => import('~/components/molecules/registermenu-info')
   },
@@ -12,21 +13,6 @@ export default {
       type: Object,
       required: false,
       default: () => {}
-    }
-  },
-  methods: {
-    /**
-     * set menu title
-     */
-    setMenuTitle (value) {
-      this.menu.name = value
-    },
-
-    /**
-     * set menu desciption
-     */
-    setMenuDescription (value) {
-      this.menu.discritpion = value
     }
   }
 }