+++ pages/menu/register/make/index.vue
@@ -1,9 +1,10 @@
 <template>
-  <editPreperation :menu="menu" @setMenuPreperation="setMenuPreperation" />
+  <editPreperation :menu="menu" />
 </template>
 
 <script>
 export default {
+  middleware: 'authenticated',
   components: {
     editPreperation: () => import('~/components/molecules/registermenu-preperation')
   },
@@ -12,14 +13,6 @@ export default {
       type: Object,
       required: false,
       default: () => {}
-    }
-  },
-  methods: {
-    /**
-     * set menu perperation
-     */
-    setMenuPreperation (value) {
-      this.menu.preperation = value
     }
   }
 }