+++ store/state.js
@@ -1,0 +1,8 @@
+/**
+ * state
+ */
+export default () => ({
+  menus: [],
+  selectedMenu: {},
+  loggedin: false
+})