+++ Changelog.md
@@ -2,7 +2,29 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+## [Unreleased]
+## [1.0.3]
+
 ## [Released]
+## [1.0.2]
+### Added
+- Add swipeable pages | 2020-04-17
+- Edit menu | 2020-04-16
+- Register page + functionality | 2020-04-19
+
+### Changed
+- Register pages added middleware auth | 2020-04-16
+- Refactor store to action,mutation,state,getters files | 2020-04-17
+- Refactor images to one place in project | 2020-04-17
+- Update code to refactor images | 2020-04-17
+- Update code to refactor store | 2020-04-17
+- Updated filter label text | 2020-04-17
+
+## Hotfix - 2020-04-14
+### Changed
+ - Add ingredient validation | 2020-04-13
+ - Add ingredient info placeholder | 2020-04-13
+ - Changed the default theme to custum theme | 2020-04-14
 
 ## [1.0.1] - 2020-04-12
 ### Added